Remove plugin and move to services
