--- plugins/nemo-jailbreak-detect/pyproject.toml
@@ -1,61 +1,0 @@
-[project]
-name = "nemo-jailbreak-detect-plugin"
-version = "0.1.0"
-description = "Self-hosted NemoGuard JailbreakDetect model deployment, decoupled from NIM."
-requires-python = ">=3.11,<3.14"
-dependencies = [
-    "nemo-platform-plugin",
-    "nemo-platform",
-    "httpx>=0.27.0",
-    "pydantic>=2.0.0",
-]
-
-# The model server's heavy runtime (torch, transformers, onnxruntime). Installed
-# into the model container, not the platform process. Kept optional so a bare
-# plugin install stays lightweight.
-[project.optional-dependencies]
-model = [
-    "torch>=2.9.0",
-    "transformers>=5.3.0",
-    "onnxruntime>=1.24.3",
-    "huggingface_hub>=1.0",
-    "einops>=0.8.2",
-    "numpy==1.26.4",
-    "fastapi>=0.103.1",
-    "uvicorn>=0.23.2",
-    "typer>=0.7.0",
-]
-
-[tool.uv.sources]
-nemo-platform = { path = "../../packages/nemo_platform", editable = true }
-nemo-platform-plugin = { path = "../../packages/nemo_platform_plugin", editable = true }
-
-[project.entry-points."nemo.services"]
-jailbreak-detect = "nemo_jailbreak_detect.service:JailbreakDetectService"
-
-[project.entry-points."nemo.controllers"]
-jailbreak-detect = "nemo_jailbreak_detect.controller:JailbreakDetectController"
-
-[project.entry-points."nemo.cli"]
-jailbreak-detect = "nemo_jailbreak_detect.cli:JailbreakDetectCLI"
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[tool.hatch.build.targets.wheel]
-packages = ["src/nemo_jailbreak_detect"]
-
-[dependency-groups]
-dev = [
-    "pytest>=8.3.4",
-    "pytest-asyncio>=0.25.3",
-    "pytest-mock>=3.14.1",
-    "ruff>=0.11.8",
-]
-
-[tool.pytest.ini_options]
-testpaths = ["tests"]
-asyncio_mode = "auto"
-pythonpath = ["src"]
-addopts = "--import-mode=importlib"
