+++ services/jailbreak-detect/deploy/deployment-config.json
@@ -1,0 +1,11 @@
+{
+  "nim_deployment": {
+    "gpu": 0,
+    "image_name": "nemo/jailbreak-detect",
+    "image_tag": "0.1.0",
+    "additional_envs": {
+      "JAILBREAK_CHECK_DEVICE": "cpu"
+    }
+  },
+  "description": "Self-hosted NemoGuard JailbreakDetect model server (CPU, runtime weight download)."
+}
