+++ services/jailbreak-detect/model/Dockerfile-GPU
@@ -1,0 +1,29 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# GPU image for the jailbreak-detection model server (pods / DGX). For local CPU
+# testing (e.g. Apple Silicon) use the CPU `Dockerfile` instead.
+#
+# Self-hosted replacement for nvcr.io/nim/nvidia/nemoguard-jailbreak-detect.
+FROM pytorch/pytorch:2.9.1-cuda12.6-cudnn9-runtime
+
+ENV JAILBREAK_CHECK_DEVICE=cuda:0
+
+# Model weights are downloaded at first start, not baked into the image:
+#   - the random forest (snowflake.onnx) into EMBEDDING_CLASSIFIER_PATH
+#   - the Snowflake embedder into HF_HOME
+# nvidia/NemoGuard-JailbreakDetect is a gated repo, so provide HF auth at RUN
+# time: `docker run -e HF_TOKEN=...`. Mount these dirs to persist downloads.
+ENV HF_HOME=/opt/nim/.cache/huggingface
+ENV EMBEDDING_CLASSIFIER_PATH=/opt/nim/.cache/jailbreak-detect
+
+WORKDIR /app
+COPY requirements.txt .
+RUN pip install --no-cache-dir -r requirements.txt
+
+# Copy only the model server sources (classifier.py + server.py).
+COPY classifier.py server.py ./
+
+EXPOSE 8000
+ENTRYPOINT ["python", "server.py", "start"]
+CMD ["--port=8000"]
