+++ services/jailbreak-detect/pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+# The model server is a standalone (non-package) container source; put model/
+# on the path so tests can `import classifier` / `import server`.
+pythonpath = model
+testpaths = tests
